🎉 Initial Cut of the Creditsafe Client
This is the initial cut of the Client with coverage for the Company and
Director endpoints. As more are needed, and time allows, they can be
added in a similar manner. All the complexities of the communication and
JWT token creation and refreshing are in the `fire()` method, so it's
just the matter of adding the endpoints, and updating the docs.