🧼  Fixed the null/indefined query params
When we passed in the argument to the search, or really, *any* of the
calls, we didn't filter out the null/undefined ones, so that when we
went to add them as query params to the call, the toString() would blow
up if they were null/undefined. This fix takes care of that so we don't
have to worry about it in the calling.